refactor: move out reading csv file
